+++ src/bigdict/__init__.py
@@ -1,7 +1,7 @@
 '''
 Bigdict
 '''
-__version__ = "0.2.5"
+__version__ = "0.2.6"
 
 from ._bigdict import Bigdict